+++ demomysql/src/main/java/xyz/hemoreg/demomysql/usermanagment/repository/ProductRepository.java
@@ -3,4 +3,4 @@ package xyz.hemoreg.demomysql.usermanagment.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import xyz.hemoreg.demomysql.usermanagment.model.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> { }
+public interface ProductRepository extends JpaRepository<Product, Long> { }
