+++ demopostgresql/src/main/java/xyz/hemoreg/demopostgresql/usermanagment/repository/ProductRepository.java
@@ -3,4 +3,4 @@ package xyz.hemoreg.demopostgresql.usermanagment.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import xyz.hemoreg.demopostgresql.usermanagment.model.Product;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> { }
+public interface ProductRepository extends JpaRepository<Product, Long> { }
