+++ demopostgresql/src/main/java/xyz/hemoreg/demopostgresql/usermanagment/service/ClientService.java
@@ -24,7 +24,7 @@ public class ClientService {
                 .collect(Collectors.toList());
     }
 
-    public ClientDTO getClientById(Integer id) {
+    public ClientDTO getClientById(Long id) {
         return clientRepository
                 .getById(id)
                 .toClientDTO();
